Add option to test jenkins node before use
Add a new node state, TEST, and if a test job name is supplied
put the node in the TEST state, and run that job with the node
name as a parameter.  If the job succeeds, move it into READY
and relabel it with the appropriate label (from the image name).

If it fails, immediately delete the node.

If it never runs, it will eventually be cleaned up by the
periodic cleanup task.

Change-Id: I5ba1ea8cdc832b13a760edaee841487afe7d7ce4