Add the ability to ignore offline targets
If a jenkins is in shutdown mode or is offline, ignore that jenkins
for the purposes of launching nodes.  Node updates (used/complete)
for that jenkins will still be processed.

This should allow another jenkins to gracefully accept the increased
load if one goes offline.

Also, log the IP address when spinning up a node.

Change-Id: I3a8720dd5aaf154ca91cdc36136decad52eb6afa