Rename ASRT -> AGT
Rename the AllocationSubRequestTarget to AllocationGrantTarget
as well as creating a new AllocationGrant class.  Currently
the AllocationSubRequest does double duty representing a request
before it's granted as well as after.  The new class takes on
the second role to make it more clear both in the code as well
as in debug output that the sub-request has transitioned from
a request to a grant.  Hopefully an AllocationGrantTarget also
makes it more clear that it holds the information about where
the nodes supplied in a grant are destined to end up.

This is not intended to produce a behavioral change other than
in debug output.

Change-Id: I8b44c35b3a99a8a3b818519a6116a74e75ced2c8