Update patch set 1
Patch Set 1: (1 inline comment)

I have a question inline about how availability_zone is used. I believe this code will work as is just not entirely in the way we anticipate (this is why I have not removed my -2). I think it is important to get this up and running and it may not be perfect on a first pass, but if anyone knows the answers to my questions inline getting answers would be great.

Patch-set: 1