Update patch set 6
Patch Set 6: I would prefer that you didn't merge this

I'm actually a bit confused.  I assume you meant 'nodepool.yaml' instead of 'nodepool.layout'.  But it also looks like the pool code isn't fully connected all the way through; I don't see an option to add it to the config file and for it to be passed through all the way to the createFloatingIP method.

I am in favor of documenting how to use the parameter, but I'm not sure the parameter exists yet.

Patch-set: 6