Update patch set 4
Patch Set 4: I would prefer that you didn't merge this

Ah, in fact this doesn't work -- the error in testing isn't an error in the test, it's a real bug.  See _getFlavors() for the structure that is being iterated over -- it's not a nova response, it's a simple list of tuples.  It probably just needs to be expanded with a third column assuming you want "find flavor with smallest amount of ram that meets the ram criteria AND substring matches the name filter".

Patch-set: 4
Label: Code-Review=-1
Label: Workflow=0