Add job runtime stats
Use state change times to determine job start and end times for a
run, and log runtime stats for the job, job+branch, job+branch+image,
job+branch+image+provider, as well as build counts for each of those.

Update the fake yaml file to align the ZMQ port with the fake servers.

Add statsd to the fake servers.

Add instructions on how to run a fake nodepool that emits statsd info.

Change-Id: I5296cafd191f363698a088fd50d5da24902fd104