Update patch set 2
Patch Set 2:

I'm not certain we'd use them, and it'll take up a quite a bit of space, so I'd prefer to just go with SUCCESS for now.  We can add the result to the end of the key in the future if we change our minds.

Patch-set: 2
Label: Code-Review=0