Support install phase with nodepool
Some operating systems are not available as cloud images. To be able to
create nodes of such operating systems, we must install them first on
top of, or by replacing the standard os. This install phase can
include several automated restarts of the node. This patch adds this
functionality to nodepool, and the ability for nodepool to track the
installation.

New configuration values:
- image.install - specifies a script, that should be executed on the
  standard os. If it's omitted, no installation will be performed.
- image.install-done-stamp - is a path on the instance, whose existence
  is periodically polled by nodepool. The installation is ready, if the
  file exists.
  If this is not specified, the original behavior is reserved - nodepool
  assumes that the host is installed right after the install script
  returned.
- image.install-poll-interval - defaults to 30, specifies the seconds to
  wait between polling for the existence of install_status_file.
- image.install-poll-count - defaults to 10, specifies the number of
  polls nodepool performs before erroring out.

Change-Id: I107c7ada43f5650bfaa81ca99732405e749d158b