Update patch set 2
Patch Set 2: (1 inline comment)

Cool.  In general I think this is good.  The only thing that strikes me as odd is the poll loop for the status file.  Why is that necessary?  Is it so that you can timeout?  If so, what about just running the install script using /usr/bin/timeout?

Patch-set: 2
Label: Code-Review=0