Support nodes with launch condition
Some special nodes when started from an image will not be instantly
available, they might be busy for a while. Nodepool should wait, until
the nodes are ready, and they signal it through touching a file in a
configured location. Nodepool will use ssh to log in to the node, and
test if the given filestamp exists.

Change-Id: I67fd9b92e838f8dfc1a4d17bef0390677e36591c