Permit using a known keypair when bootstrapping.
For debugging it is really inconvenient to have to wait for puppet to
*work* to log into a box and analyse what went wrong. This small patch
makes making one-time keypairs optional. Setting provider.keypair to
the name of a nova keypair will use that keypair in provisioning - but
you'll need to have the real key available via an agent for nodepool
to login itself.

Change-Id: I768ba656368ff00f9f7374ec0a46e57cd563ee61