Don't load system host keys.
There are three places that nodepool connects to hosts:
 - to bootstrap where by definition any existing host key must be wrong
 - to bring up a snapshot where again any existing host key must be wrong
 - to check a node is still running

Nodepool itself doesn't save the host key, so the third case will
typically warn as well, or if an administrator logged into a node from
the nodepool user account cause a failure.

As such it seems that system host key validation is not useful and
just serves to cause failures when a stale key is present.

Change-Id: I074314d61b5c520bdd67da24bfd2a8e626da0dc5