Make cleanupServer optionally nonblocking.
cleanupServer is used from both the CLI and the daemon, but in the
daemon we don't want to block - a single slow to delete server should
not block deleting all other servers. Instead we want to pass over all
the servers we have pending deletion, submitting them for deletion,
and after that loop back and check that they were in fact deleted. Or
not.

This change also fixes a minor bug where calling cleanupServer on a
server that was concurrently deleted from the provider (e.g. by the
CLI) would result in a traceback and deferral to the next loop of
periodic cleanup before the node record was deleted.

Change-Id: If40930238d82c1103b83fe7f46ef6f6e86efd624