Default to a ratelimit of 10/second for API calls
We make a lot of API calls when deleting servers - 8 in fact to
enumerate and delete floating ips, keypairs and the server. A 1 second
rate limit between calls makes each delete take 8 seconds to move
past, and then if we're deleting 300 servers we won't get through
submitting the deletes for all three hundred for 2400 seconds, or 40
minutes.

Change-Id: I4ce86c38030e63db3547043134940fb9cf01eafa