Use the nonblocking cleanupServer.
In this patch we now change from doing one server deletion entirely at
a time to instead doing all the deletion requests and then all the
was-it-deleted checks. This gives the cloud all our requested work
at once.

Note that we still do lookups from cleanupServer, future work might
tune this more to minimise them. The reason we need to do a scan after
submitting all the deletions is to avoid waiting the whole poll period
between periodic cleanups before we notice a not-quite-immediate
delete actually happened.

Change-Id: I8f52b24ed911c1e6b8e04b8fe561100f8552fee6