Update patch set 6
Patch Set 6:

Bob, thanks for updating the patch. I think we will need one additional paremter: a configurable sleep before we snapshot the node. This way on the end of our script, we could shut down the instance, and the sleep would give the node some time to shut down the hypervisor.

Patch-set: 6
Label: Code-Review=0