Update patch set 7
Patch Set 7:

In my opinion, this change adds the install phase, and the second patch adds the status poll for node readiness - these are separate things. On the second change, I am using the same polling mechanism, as here, so that's why a small refactor is part of that change to avoid code duplication. So the next change is not purely a refactor. For more information:

https://wiki.openstack.org/wiki/XenServer/GateIntegration

Patch-set: 7
Reviewer: Gerrit User 5044 <5044@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0