Adding a cleanup timeout to nodepool
In big projects like openstack, who has > 400 VM's running,
the default cleanup timeout (8 hours) is fine. However, for
small projects with 5 or 10 VM's running, it's a waste
of resources keep an image for 8 hours in an error state.
This patch adds the possibility to  pass a timeout
to wait until an image in error state can be removed.

This is a retake on previously abandoned change:
https://review.openstack.org/#/c/71261/

Change-Id: Ic8d63484d626d19f5da56f7a8cbc39a58ebac6af