Adding the possibility to pass a cleanup timeout to nodepool
In big projects like openstack, who has > 400 VM's running,
the default cleanup timeout (8 hours) is fine. However, for
small projects with like 5 or 10 VM's running, it's a waste
of resources keep an image for 8 hours in an error state.
This patch adds the possibility to the projects pass how long
they want to wait untill remove an image in error state.

Change-Id: I1070ea59244a9844feb96729679aef74b5058b1a