Update patch set 9
Patch Set 9: Abandoned

Of course, after talking to Mate, this now makes sense.
We were running nodepool using sudo from the ubuntu account - my patch is now needed if the key is in ~ubuntu/.ssh but it just works if the key is in ~root/.ssh.

I'm guessing this is because paramiko would previously have cached the key, but without that caching I got it working by loading it through the image.private_key.

This isn't needed - abandoning.  Thanks for the skepticism lifeless!

Patch-set: 9