Support offline providers on startup
If an error occurred while initializing a provider, don't add it to the
list of providers. This will cause it to retry the next time the main
loop is executed. Previously if an error occured during startup the list
of providers being iterated wouldn't get fully processed.

There was a workaround for this in ProviderManager._getFlavors but a
similar workaround was also needed in _getExtensions. To handle the more
general case better I've removed this workaround.

Finally if a image can't be updated because its provider hasn't been
added, remove it from the database so it will retry once the provider
becomes available.

Change-Id: I13e98deea10e040109fffe7280ae482aec2e5352