Catch key problems in ssh_connect
paramiko throws PasswordRequiredException if they key being used isn't
added to the users authorized_keys. This can happen if we tried to ssh
to a host before cloud-init has configured ssh. To avoid creating a
new server each time this happens we catch this error and treat
it similar to connection refused.

ssh_connect will now continue to retry until root's key is configured.
In the case of the ubuntu cloud image will

Change-Id: I33ff8c0ddaae903cc427437508928177d867358b