Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

(1 inline comment)

There's an inline suggestion (the -1 is for that, what follows is for discussion), but testing this made me realize that the problem may be a little more basic.  This does prevent images from being created for offline providers, but the main loop is still significantly slowed.  Each time through it's going to try to connect to any offline providers which will cause it to pause for 60 seconds each.  That's probably survivable, but it's not great; it's a pretty significant performance degredation (the main loop normally runs every ~13 seconds in production).

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1
Label: Workflow=0