Reset the client object before creating the image
In situations where the connection between nodepool and the cloud
it manages may be through an evil proxy, the fact that the
requests library in python-novaclient keeps port connections open
means that the evil proxy might timeout the connection during the
period of time that the image is being built.

Resetting the connection before calling create_image will ensure
that the client object is not stale.

Change-Id: I3d69de48be1993c04356d209f622b0a17e6e8d03