Reuse floating ips
A floating ip doesn't carry state between uses, and doesn't really
need to be flushed. Since we're keeping an upper cap of the total
number of nodes in a pool anyway, it's actually possible to keep a
set of floating ips that are known and re-used. This means that it's
possible to either pre-create the a set of them equal to the size
of the quota and then list them somewhere else, such as iptables
rules, so that a known quantity of source ips can have access to
things.

To achieve this, stop deleting floating ips at the end of a server's
life, and try first to re-use any floating ips that are not yet
attached to a server.

Change-Id: I5e64b85a75b518b33f53f0c41bce6e09106a8df2