Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

(1 inline comment)

One small suggestion to make the code cleaner inline.

I feel like this is the sort of change that is dangerous because it seems so similar and that is where the dragons live. But other than the inline thing this lgtm.

Patch-set: 2
Label: Code-Review=-1
Label: Workflow=0