Synchronize floating IP ops to prevent leaks.
Floating IP deletion may not always be able to associate an IP address
to a node being deleted. To avoid leaks in the case delete any floating
IP that is not associated to a node. This introduces a race between
creation of a floating IP and association of an IP to a node where an IP
may be improperly deleted. Synchronize these operations using a lock to
prevent this race and make the more zealous deletes correct.

Closes-Bug: #1301111
Change-Id: I2aefde0665896698a8ad76f6f423f55b8728b923