Shuffle requests for fairer label allocation
When grants are becoming available slowly one at a time (e.g. when
max-servers has been hit for a provider), the allocation algorithm
needs to shuffle the list of requests in order not to favour the
first label type in the list on all iterations.

This adds a weighted shuffle to the list of reqs in order to
preference labels based on the ratio they were configured with.

Also because in this case only a single allocation is available
and the ratio is low, grant will always be set to 0. If this
occurs we set grant to 1, the weighted shuffle will ensure
this only happens as often as is appropriate.

Change-Id: Iffd1e973ca5a385446519f1059b285511191fa2e
Fixes-Bug: #1308407