Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

I share Jeremy's concern though I'm having a little bit of a problem identifying when it might be a concern.

Also, there is a race condition with Jenkins possibly starting a job on a node that nodepool has just started deleting.  I think if we wanted to do this, we would need to be very careful about it; possibly doing some kind of two-phase process where the node is relabled, and then we wait to make sure no job has started, and then delete the node.

However, I think there are circumstances where our demand might change in such a way as to make this behave poorly for us.  For instance, if a new revision is uploaded that has a merge conflict, we will have canceled jobs in the queue and not replaced them, resulting in a net loss of nodes (and in fact, as Jeremy points out, they would be deleted immediately after being built).  I don't have numbers, but I expect our system is busy enough that it happens occasionally, and we would be at risk of unecessary node deletions (and the above race condition).

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1
Label: Workflow=0