Update patch set 2
Patch Set 2:

That's right - the pool of flooating IPs would not go down. I think being able to put in a list of known IPs into iptables (which might be nice once we've got direct-gearman-workers - even with ssl certs)

We could make it configuable if you think that we might want to explicitly scale up and down, but it seems given how things work for us, less churn up and down might be friendlier to our clouds.

Patch-set: 2