Support provider AZ lists
Apparently not all clouds will schedule AZs as expected by nova. It may
be the case that an AZ is hard set on the provider side when no specific
AZ is requested. Add AZ support to nodepool so that it can request AZs
and better load balance across AZs provided.

Do this by randomly selecting an AZ from the list of AZs provided in the
config. This should give us a good distribution across all AZs. If a
differently weighted distribution is required a nodepool provider object
can be created per AZ with a single item AZ list.

Note this requires an update to the nodepool database.

Change-Id: I428336ad817a8eb7d311a68767849aab0bcf015f