Remove the need for zmq in proper demand calcs.
Where zmq is not used and nodes are never marked as in use nodepool
was unable to predict how large to grow the pool in advanced. It
wouldn't launch new nodes until the current ones start to be deleted.

Previously the demand was calculated as:
start_demand = queue - running_jobs
min_demand = start_demand + min_ready
demand = min_demand - (runnning + building)

If you make an assumption that a nodepool node will be used to do
a job for the gearman it is connected to it becomes possible to
calculate the actual demand off just the gearman statistics instead.

As such, count the total jobs in the queue as the demand (inclusive
of running jobs). Then minus off nodes considered to be USED. These
therefore account for the running jobs and lower the demand since
they aren't actually waiting for a node. This give us:

start_demand = queue
min_demand = start_demand + min_ready
demand = min_demand - (running + building + used)

So assuming used == running_jobs (which it should in theory if there
are no foreign workers) then you can see the above calculations
are equal.

Removing the requirement of zmq makes it easier for other workers,
such as turbo-hipster, to be launched by nodepool without the need
of sending back an event stream.

Change-Id: I23089513ae4ff58388cfa82030b21f2ed41e51b9