Update patch set 2
Patch Set 2:

It may be the case that we are changing the hostname to match the instance name. In fact, the current prepare_node.sh scripts in openstack-infra/config expect the hostname as a parameter and use the `hostname` command to set the value appropriately.  As a fluke, because this expected functionality isn't in place the prepare_node.sh script simply prints the current hostname because the value of $1 is unset.

Patch-set: 2
Reviewer: Gerrit User 6039 <6039@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0