Grant at least one allocation
When over quota, we can start to drop small allocations completely.
This can stop forward progress -- if that node is required to complete
a gate job, no gate jobs can pass until that node runs the job;
however in the mean time, more gate requests are piling up making the
over-quota situation even worse.

This was seen to happen with the Fedora gate job for devstack.
Although we want a comparatively small number of Fedora nodes to
process a devstack change, they were never created due to the
extremely high load of other node types.

This adds a flag to ensure at least one of each type is created if
requested.

Change-Id: I090ef256b6f0b12d2999fc7cae597b911d06ce38