[RFC] Track last allocations to ensure forward-progress
Track the last round of allocations to ensure that we don't starve a
particular label.  If a label made a request but didn't get any nodes
the last time round, it is given nodes preferentially on the next
calculation.

(no test cases yet, looking for comments on mechanism before I
implement them)

Change-Id: I0ff9aa74fef807bd84bf51e7ba1ed176c22f5365