Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

I know nodepool is generally very reliable but we have had issues with it from time to time - and I've been considering how to address that. Running two independent ones would do that - but this change would (AFAICT) make that impossible. I may be wrong, or it may not be how infra-core would like to see that tackled. If so - my remaining ask is for a test ;).

Patch-set: 2
Label: Code-Review=-1