Update patch set 2
Patch Set 2: Workflow-1

Yeah, I think you are spot on, Robert. I'll have a think about it. At this point though I think I may end up abandoning this patch and figuring out a way for other workers to communicate with nodepool (which may be zmq as it is). It would, however, be nice if nodepool knew what to do from gearman alone.

Patch-set: 2
Label: Workflow=-1