Update patch set 3
Patch Set 3:

The ready script is used in the "launch" phase, not the template phase.

The above script is needed for before we take the snapshot so can't use the ready script.

Possibly the other change (launch_done_stamp) could be re-written to 'ready_done_stamp" although XS does not need a specific 'ready' script after booting the VM.

Patch-set: 3
Label: Verified=0