Update patch set 22
Patch Set 22: Code-Review-1

(4 comments)

I think the general direction is good, but the code in _updateImage illustrates where we're going to need to do some reworking.

I think it may be too complicated to keep both methods of building an image -- I'd be happy for the old one to be removed if that helps.

Here are the cases I think we need to handle:

checkForMissingImages:
  do the following in the background (probably in a thread, but there should only be one of these threads at a time):
    if any label images are missing, build them
    if any providers are missing images, upload them

periodic image update:
  (doesn't need to be its own thread)
  update each label image
  upload to all providers

command line:
  new 'image-build' command to run dib and produce a new image file
  new 'image-upload' command to upload it to a single provider
    argument to upload it to all providers
  make 'image-update' do both of those (for all providers)

Because building an image is slow, and we don't want to do more than one at a time locally, and uploading is (relatively) fast, that will probably change how some of the threading is done.  And you need to make sure that checkForMissingImages and the periodic image updates don't try to build images at the same time (maybe with a lock, or maybe with a consumer queue thread that builds images).

Patch-set: 22
Label: Code-Review=-1