Record nodeinfo as ansible facts and inventory
Ansible has the ability to automatically load local 'facts' about
a system from files. Emit the nodepool multi-node info also in
this manner so that we can potentially key off of it from ansible.
Also, this writes an ansible inventory file on each host such that
if a multi-node job wanted to be written in ansible, appropriate groups
would be set up.

Change-Id: Ica52b4b47c1d1fa61fc66e4f2cb50b99dc15ac9c