Record nodeinfo as ansible facts and inventory
Ansible has the ability to automatically load local 'facts' about
a system from files, and supports inventory files. Record all of the
multi-node info into it.

The format used here is the same as what is currently used in
devstack-gate - but this would expand that availability of the multinode
ansible structure to non-devstack things too.

Change-Id: Ica52b4b47c1d1fa61fc66e4f2cb50b99dc15ac9c