Filter negative available count for allocator
The issue described in [1] describes a large over-allocation of nodes;
greater than the max-servers value set for a provider.

Looking at the code in AllocationProvider.makeGrants(), it does not
seem there is a path where a grant can be made for more than
"self.available".  I do not think this is the failure path.

However, if the available nodes is set negative, the calculation
algorithms start subtracting a negative value (i.e. adding) and we can
see strange results.  For example, without the fixes, the test-case
sample is requesting 10 nodes, but ends up with 7 allocations of
label1 and 8 of label2 (i.e. 5 more nodes than it asked for!).

It seems a negative allocation could happen.  The available nodes are
calulated in nodepool.py by

 n_provider = count_nodes_and_subnodes(provider.name)
 available = provider_max - n_provider

where count_nodes_and_subnodes() is querying the database to find the
active nodes.  I can not explain how this might have become
out-of-sync, but if it were to return more than max-servers for the
provider we would end up in this case.  Assuming too many nodes were
spun-up, this would make things worse; the result of
count_nodes_and_subnodes() gets larger and larger, which means
available falls further negative and even more nodes get erroneously
created.

Clamping this to a minimum of zero should avoid this distructive
situation.

Also added is a error log to highlight when the nodepool has found
this situation.  An examination of the database v actual running might
help diagnose the root cause if this error is seen.

[1] https://review.openstack.org/#/c/109132/

Change-Id: I40f432b00e619e38135d34ae15c98f12df52424b