Update patch set 1
Patch Set 1: Code-Review+2

Not approving because I would like jeblair to indicate that we are ok with giving up for now on yaml verification. (I am ok with it because it hasn't happened and this dep can be added when it does happen).

Patch-set: 1
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2