Update patch set 2
Patch Set 2: Code-Review-1

I agree that this change is reasonable.  However, it's being proposed to fix a problem that we saw with the history-tracking allocator, and we have not tracked down the root cause of that.  Because this papers over the problem, merging this at the moment would inhibit our ability to find the root cause.  I'd like to thoroughly fix that first, and then merge this as a belt-and-braces fix.

Patch-set: 2
Label: Code-Review=-1