Log error on over-allocation
I have a suspicion that the reason the history-tracking changes
started going crazy was that there was somehow a discrepancy between
max-servers and the results of session.getNodes().

If this were to happen, a negative available value could be passed to
the allocator which causes it to go crazy.  If this happens, log an
error.

Change-Id: Ibd63bbdbbfc7868057872b97b159e8481c3efab2