Remove unneeded call to removeFloatingIP in cleanupServer
When cleaning up a server previously a call was made to removeFloatingIP()
which disassociates the floating ip from the server. Next, a call is made
to delete the floating ip. Nova allows us to delete the floating ip directly
even if it's associated with a server so there is no need for us to first
unassociate it.

Change-Id: I0c78823198fac0d31235d93505a4251edbf9e612