Update patch set 2
Patch Set 2: Code-Review+1

Not +2 because I am on the fence about this. It won't actually fix the problem aiui. The same leak is there using novaclient because it does the same disassociate the delete that we do.

So really this is just a code simplification. It may make this easier to debug in the future but not correct the leak.

Patch-set: 2
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1