Update patch set 2
Patch Set 2:

Hi Clark, 

"Not +2 because I am on the fence about this. It won't actually fix the problem aiui. The same leak is there using novaclient because it does the same disassociate the delete that we do."

You mean nova-api not the nova-client right? The nova-api fix is here: https://review.openstack.org/#/c/113754/ . You are correct that someone needs this nova-api fix in place for the leak to stop unless we want to make nodepool keep up with the floatingip_id's or use the neutronclient to delete the fip though i think it would be better to bite that integration chunk off at a different time.

Patch-set: 2
Reviewer: Gerrit User 4395 <4395@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0