Refactor nodepool to configure glance only when uploading imaes
Currently glance was configured on nodepool start, when
reconfiguring provider managers. This caused two issues:
1. If the cloud was down, nodepool couldn't configure the manager
due to an exception and was stuck on reconfiguring forever
2. Normally keystone auth tokens have an expiration of 1 hour.
If configured only once, auth token would expire causing
glance connection to be invalid.

Configuring glance only when uploading images would solve
that problem, and there is no need to have it always active,
as image uploads will be happening normally once a day.

Change-Id: I0b61f576b265c705408d1e483fc5f5f8a33d3ee3