Update patch set 1
Patch Set 1:

Good for changing that, to allow more flexibility. But now that we are in that direction, can we suppose that jenkins will be installed always in username home directory? Maybe we could add a setting for the jenkins path in the config file, and if empty, default to /home/username?

Patch-set: 1
Label: Verified=0