Add meta option for passing meta-data
Add a meta option for passing through metadata key/value pairs.  For
snapshot images, this ends up getting passed through to
novaclient.server.create_image(metadata=...).  For uploaded images, the
values are used as custom properties for glanceclient.images.create()

The 255 character limits are taken from the API documentation in [1].

This is initially intended to be used for setting xenapi and vm_mode
settings for disk-image-builder based Rackspace images, as described
in [2].

[1] http://docs.openstack.org/developer/python-novaclient/ref/v1_1/servers.html
[2] https://developer.rackspace.com/blog/custom-images-via-boot-dot-rackspace-dot-com-training-wheels-included/

Change-Id: Ic2f044f44f06830b5b84438d838013d409e642b4