Allow passing cache directory to disk-image-create
When building images using dib, the disk-image-create command
is defaulting to ~/.cache/image-create as cache directory. That
causes problems with disk free space, as this operation is quite
space consuming. Setting the cache-dir parameter forces
disk-image-create to use the specified directory to store cache,
allowing to point to some volume with extra disk space.

Change-Id: I1cc5761f0c2de4724dc6a86cd44dd98a10495790