Add alembic migration support
When we make changes to the database schema, currently we have to
manually apply the schema updates. This is not only ugly for us, it's
ugly for anyone who happens to be tracking nodepool. By adding alembic,
we can update our automation to run nodepool-db-manage upgrade head
every time we install a new version of nodepool, and thus ensure that
database changes are automatic.

Change-Id: Ie7d6276c7c975489106c089f6aa7446d347827e4