Add alembic migration support
When we make changes to the database schema, currently we have to
manually apply the schema updates. This is not only ugly for us, it's
ugly for anyone who happens to be tracking nodepool. By adding alembic,
we can update our automation to run nodepool-db-manage upgrade head
every time we install a new version of nodepool, and thus ensure that
database changes are automatic.

Three distinct modes are handled here:

1) Database is empty.

   The Database will be populated with the current schema directly
   and stamped as "up-to-date".

2) Database is populated and under alembic management.

   The migrations will automatically be run on startup ensuring that
   the db will be in a consistent state for the code.

3) Database is populated and not under alembic management.

   The database is assumed to be "current" and is simply stamped as
   being "up-to-date". It is safest to simply drop the database and
   let node pool repopulate it as per previous schema changes when
   initially upgrading from a non-alembic enabled version of
   nodepool to an alembic version of nodepool.

Change-Id: Ie7d6276c7c975489106c089f6aa7446d347827e4