Remove unnecessary path joining for dib filenames
We store the dib filename path as os.path.join(imagesdir, filename) in
the database. Any code that needs the full path to the file that gets
the filename from the database does not need to join the path to the
filename.

Thankfully this all worked because os.path.join is smart when dealing
with rooted paths so os.path.join('/foo/bar', '/foo/bar/baz') returns
'/foo/bar/baz'. But this should only work with fully rooted paths so
make these changes on the off chance that someone wants a relative path
for some reason.

Change-Id: I769c47fa8089668371cfb28c84f6bde7760cb504