Handle numeric "release"
If the "release" is numeric, such as for Fedora, it gets interpreted
as an integer and blows up when we try to insert it into the
environment (see I0d4e8519972d14fc100f46dae1b4549743b438f2).  Although
you can type the yaml input, automatically convert it to a string to
follow principle of least-surprise.

Test-case is updated.  The whole image is made to look "fedora-like"
to avoid confusion.

Change-Id: I0ea4dc54edad7ef9b5d4ad59922f762ed8bbcd89