Make update image understand both types of images
Prior to this change the image-update command did not understand that
it shouldn't always build and upload a dib image if a label is provided
by dib images and snapshots. Teach it how to build dib images and upload
them only when the provider:label pair is built this way.

Change-Id: I067b290919e3c744a4d39326691315532a7e6615