Update patch set 1
Patch Set 1: Code-Review-1

Two things. Can we fix the commit message so that there is a title line followed by a blank line followed by the rest of the message? makes it much easier to read parse for many things.

And second we should specifically exclude the broken 2.0.x versions, so that only 2.0.4 and greater is allowed.

Patch-set: 1
Label: Code-Review=-1