Better image checking in update_image command
Previously we were not handling the case where label.name != label.image
or cases where multiple labels have the same image but these are
perfectly valid cases for the update_image command. Deal with this by
iterating over all labels, any that have an image that matches the
requests image update have their providers added to a dict that is
iterated over later to ensure every provider providing the image for a
valid label is covered.

Change-Id: I0b9f04922c90f52ced4d4a0b03e20a1b20fa7f8a