Update patch set 1
Patch Set 1: Code-Review+2

(1 comment)

Thanks for debugging this. One comment inline with a question about loggin.

The only other thing is that the code makes excessive use of continues. I always prefer iterating over a complete set of data that is actionable but I can't think of a way to do that here that remains as readable as the proposed change. We would need to keep a couple dimensions of data for mapping providers and images in use onto snapshots or diskimages. Probably not worth changing.

Patch-set: 1
Label: Code-Review=+2