Add generate-log-config command
Add a command 'generate-log-config' to the command-line interface that
reads a nodepool configuration and outputs a sensible logging
configuration that separates out the general logging and image-build
logging into separate files.

While the main intent of this is to create a logging configuration
file suitable for use in upstream openstack servers, the output file
is suitable for general purpose use.

A test-case that parses a generated configuration file (based on a
real upstream sample configuration) through logging.confg.fileConfig()
is included.  Documentation is updated.

The documentation for logging.config.fileConfig() makes it seem likely
that with some minor modification nodepool can re-read this file just
as it re-reads its config file.  This is left to a later change.
puppet changes to use this command to deploy the configuration file
are in I90af348cea41e95fe0795bfef618073d4aabb1c2.

Change-Id: I916f3dd9634ca7ac691bef98236f6cf549fb7059