Add generate-log-config command
Add a command 'generate-log-config' to the command-line interface that
reads a nodepool configuration and outputs a sensible logging
configuration that separates out the general logging and image-build
logging into separate files.

While the main intent of this is to create a logging configuration
file suitable for use in upstream openstack servers, the output file
is suitable for general purpose use.

A test-case that parses a generated configuration file (based on a
real upstream sample configuration) through logging.confg.fileConfig()
is included.

Change-Id: I916f3dd9634ca7ac691bef98236f6cf549fb7059