Update patch set 7
Patch Set 7:

Not scoring this review yet because I think I need a bit more info. Overall I like the change. I think the math is sound, it is well documented, and overall makes demand calculation a lot easier to follow.

Where I get a bit lost is this relies on checking the database for nodes in a USED state instead of checking gearman for number of running jobs. The thing that updates the database rows to set node state to USED is the ZMQ event listener. When an event comes in that says job started on node X the databse row for node X is updated to change the state to USED.

So does this change actually remove the ZMQ dependency for node calculation? Even if it doesn't I think it may be a good change to have simply because it has made the code much more readable, but we should remove the bits about ZMQ if I am reading this correctly. And if I am wrong even better :)

To address the HA nodepool questions I have actually thought it would be better to run a nodepool per cloud in that cloud. Most of our reliability issues seem to come from clouds themselves, nodepool doesn't really go down all that much. So by partitioning ourselves off like that if cloud goes away so does the nodepool working that cloud and no big deal.

Patch-set: 7
Label: Verified=0