Add nodepool command tests
This commit does a small refactor of the base DB test class so that it
can be used by other child classes without duplicating the work in the
existing nodepool test cases. With this in place add a new test class to
test nodepool commands.

Currently only image-update is tested, but it should be easy to add
tests for the other commands that nodepool supports.

Change-Id: I6ea488a896cc456600da576af6086f95f4938bde