WIP: Not even close to finished - use shade
Replace all direct calls to python-*client with calls to shade. When
this is done, there should be no openstack logic workarounds in
nodepool.py, they should all be contained in provider_manager. Also -
"Business logic" about how to deal with the differences in clouds should
be in shade. Logic about how nodepool wants to use it should be in
nodepool.

Tests passing, need to discuss pending calls and really test it.

Change-Id: I7caeac098334fcce079af8e5078752243a5698e8