Update patch set 1
Patch Set 1: Code-Review-1

Overall this is good. I have a slight perference for specifically recording an ipv4 address and an ipv6 address, but the current DB schema doesn't really accomodate this so that work will require a DB schema update. Due to this I think we can start here and add on the ipv4 and ipv6 DB columns in a separate change. Something to think about though.

Ok now for the -1. It would be really good to have a test for this. I don't think it would be too hard to provide if you update the fake-provider to return ipv6 and ipv4 addrs then let the ipv6 preferred machinery switch which one you get back as the node ip. Start by adding a v6 address to the address list returned by fakeprovider.py's create method, and add a test in test_nodepool.py.

Patch-set: 1
Label: Code-Review=-1