Update patch set 7
Patch Set 7:

Good catch that the USED note is coming originally from zmq. However the point of this change is that there becomes no difference between READY and USED.

Previously the start_demand was just what gearman thought wasn't in build state. The total demand for the label was given by the min_demand minus nodes that are ready to take jobs.

Now the start_demand is all the jobs (building or not). And the total demand is given but the min_demand minus all nodes whether they are building, ready or used.

The point here is that if USED isn't updated from zmq then anything in USED will be in READY. Since these are summed at this point it doesn't matter if nodepool knows which nodes are which.

So where zmq is implemented the nodepool page will report which nodes are being USED. Where zmq is disabled nodepool will list all nodes as READY even if one may have a node (which is how turbo-hipster works).

I hope that helps, I don't feel like I've articulated it very well so let me know if I need to go again. I'll update the commit message to fix the terminology a little.

Patch-set: 7
Label: Verified=0