Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

this doesn't really work ... os.getlogin() gets you the uid, but not the euid so if i sudo to nodepool this will get it wrong.

in general i just don't see the need for this.  i guess the intent is to stop you accidentally running nodepool as root and it creating things as root that then fail when you go back to a script that runs it as nodepool.

Patch-set: 2
Label: Code-Review=-1