Update patch set 4
Patch Set 4: Code-Review-1

This assumes that everyone running nodepool has deployed it with a user 'nodepool'. I'm not sure if this is a good assumption or not.

Perhaps we don't want to set a default for the --enforced-user but rather only enforce when that option is passed or maybe based off the config?

Patch-set: 4
Reviewer: Gerrit User 7069 <7069@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1