Update patch set 4
Patch Set 4: Code-Review-1

honestly i'm not really on board with the motivation for this change.

if you're running as a peer user, then presumably you can't overwrite anything anyway thanks to permissions

so what this is *really* checking is if you are root or not; in that case I can see you can overwrite things and create permissions errors later on.

but ... you can do this with lots of tools and indeed that really is the point of being root anyway.  if every tool said "don't run this as root" and made you specify some different flag we'd probably all go insane

Patch-set: 4
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1