Update patch set 4
Patch Set 4: -Code-Review

Perhaps we could change the logic from "Don't run unless you are --enforced-user" to "Set the user to --nodepool-user" when running this?

The value could be set via config, be a parameter w/ 'nodepool' default, etc.

I get the arguments against the patch, but there also seems to be enough of an irritant around this to try to get a fix.

Patch-set: 4