Pretty Age in nodepool list
This changes the Age column in the nodepool list output to the format HH:MM:SS
The time difference is rounded to a whole number of seconds.  Further precision
is not needed.

For a time difference greater than 24 hours, the output may be like:
1 day, 0:52:11

This aims to be both human and machine readable.

Change-Id: I9ab1bc1cadedbd64522cbe72d6004f5e95c98db1