Pretty Age in nodepool list
This creates new column in 'nodepool list' output with a "pretty"
format like "1m 03s".  I find this more readable than hundredths of hours.
The addition of a column attempts to preserve backwards compatibility
with any existing machine processing of this output.

The code is similar to how Zuul generates status messages for Gerrit.

Change-Id: I9ab1bc1cadedbd64522cbe72d6004f5e95c98db1