Test that dib failures don't stop node builds
Test that when a dib image fails to build nodepool continues to build
nodes using images it does have available.

To check that the dib build actually fails update the diskimage-create
fake script to touch a file that can be checked in the TMPDIR if the
TMPDIR isn't set properly.

Change-Id: Ie7c13af319141ec3cf8d43155358157e7d9274b8