Test that dib failures don't stop node builds
Test that when a dib image fails to build nodepool continues to build
nodes using images it does have available.

To test that dib actually fails check the logged subprocesses for the
tests for exit status of 127. This is the value that is returned when
the fake dib builds are explicitly told to fail.

Change-Id: Ie7c13af319141ec3cf8d43155358157e7d9274b8