Update patch set 4
Patch Set 4: Code-Review-1

I'm pretty sure we are double counting ready when calculating extra_demand. As an example, if we have start_demand=5, ready=1, and min_ready=3, then we should end up with a demand of either 4 or 7 depending on capacity to give us a total of either 5 or 8 nodes. But this will give us demand=4 and extra_demand=2, for a total demand of either 4 or 6. If we have capacity for both start_demand and min_ready, then we end up with a total of start_demand + min_ready - ready.

Patch-set: 4
Label: Code-Review=-1